Adicionando suite testeFuncionalidades
